PSAP-1354: Updated a script for deploying a custom NTO image to a cluster (#1030)
* Updated a script to deploy a custom NTO image to a cluster

Using this script developers could easily test their code.
Running it will replace the NTO image in the current running operator deployment with the custom image.

* Updated HACKING.md to build & deploy custom NTO image

It is important to make sure this file contains correct steps, and does not turn outdated.
Therefore, a new contributor reading HACKING.md file is able to deploy a custom version of NTO.